+++ kpi/views/v1/export_task.py
@@ -6,10 +6,9 @@ from rest_framework.response import Response
 from rest_framework.reverse import reverse
 
 from kpi.models import ExportTask
-from kpi.models.import_export_task import _resolve_url_to_asset
 from kpi.serializers import ExportTaskSerializer
 from kpi.tasks import export_in_background
-from kpi.utils.models import remove_string_prefix
+from kpi.utils.models import remove_string_prefix, resolve_url_to_asset
 from kpi.views.no_update_model import NoUpdateModelViewSet
 
 
@@ -194,7 +193,7 @@ class ExportTaskViewSet(NoUpdateModelViewSet):
             raise serializers.ValidationError(
                 {'source': 'This field is required.'})
         # Get the source object
-        source = _resolve_url_to_asset(
+        source = resolve_url_to_asset(
             task_data['source'])
         # Complain if it's not deployed
         if not source.has_deployment:
